requirements.txt
alembic==1.13.2
click==8.1.7
Flask==2.0.3
Flask-Continuum==0.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==0.19.2
SQLAlchemy==1.4.15
typing_extensions==4.12.2
Werkzeug==2.3.0
